+++ public/js/plugins.js
@@ -28,6 +28,13 @@ export const officialPlugins = [
     kind: 'editor',
   },
   {
+    name: 'Subscriber Later Binding (GOOSE)',
+    src: '/src/editors/GooseSubscriberLaterBinding.js',
+    icon: 'link',
+    default: false,
+    kind: 'editor',
+  },
+  {
     name: 'Subscriber (SMV)',
     src: '/src/editors/SampledValuesSubscription.js',
     icon: 'link',
@@ -38,7 +45,7 @@ export const officialPlugins = [
     name: 'Subscriber Later Binding (SMV)',
     src: '/src/editors/SMVSubscriberLaterBinding.js',
     icon: 'link',
-    default: false,
+    default: true,
     kind: 'editor',
   },
   {
@@ -185,6 +192,15 @@ export const officialPlugins = [
     kind: 'menu',
     requireDoc: true,
     position: 'middle',
+  },
+  {
+    name: 'Show SCL History',
+    src: '/src/menu/SclHistory.js',
+    icon: 'history_toggle_off',
+    default: true,
+    kind: 'menu',
+    requireDoc: true,
+    position: 'bottom',
   },
   {
     name: 'Help',