+++ src/editors/SMVSubscriberLaterBinding.ts
@@ -1,6 +1,7 @@
 import { css, html, LitElement, property, TemplateResult } from 'lit-element';
 
-import './subscription/smv-laterbinding/svc-laterbinding-list.js';
+import './subscription/later-binding/fcda-later-binding-list.js';
+import './subscription/smv-laterbinding/ext-ref-laterbinding-list.js';
 
 /** An editor [[`plugin`]] for Subscribe Later Binding (SMV). */
 export default class SMVSubscribeLaterBindingPlugin extends LitElement {
@@ -10,10 +11,14 @@ export default class SMVSubscribeLaterBindingPlugin extends LitElement {
   render(): TemplateResult {
     return html`<div>
       <div class="container">
-        <svc-later-binding-list
+        <fcda-later-binding-list
           class="column"
           .doc=${this.doc}
-        ></svc-later-binding-list>
+          controlTag="SampledValueControl"
+        >
+        </fcda-later-binding-list>
+        <extref-later-binding-list class="column" .doc=${this.doc}>
+        </extref-later-binding-list>
       </div>
     </div>`;
   }
@@ -24,7 +29,9 @@ export default class SMVSubscribeLaterBindingPlugin extends LitElement {
     }
 
     .container {
+      display: flex;
       padding: 8px 6px 16px;
+      height: calc(100vh - 136px);
     }
 
     .column {
@@ -32,7 +39,7 @@ export default class SMVSubscribeLaterBindingPlugin extends LitElement {
       margin: 0px 6px 0px;
       min-width: 300px;
       height: 100%;
-      overflow-y: scroll;
+      overflow-y: auto;
     }
   `;
 }