+++ src/foundation/compare.ts
@@ -156,8 +156,8 @@ export function diffSclAttributes(
       )
     );
     if (
-      !shouldFilter &&
-      elementToCompareAgainst.getAttribute(name) !==
+        !shouldFilter &&
+        elementToCompareAgainst.getAttribute(name) !==
         elementToBeCompared.getAttribute(name)
     ) {
       attrDiffs.push([
@@ -277,9 +277,23 @@ export function diffSclChilds(
 export function renderDiff(
   elementToBeCompared: Element,
   elementToCompareAgainst: Element,
+  filterToIgnore: DiffFilter<Element> = {}
+): TemplateResult | null {
+  return renderDiffInternal(
+    elementToBeCompared,
+    elementToCompareAgainst,
+    filterToIgnore,
+    elementToBeCompared,
+    elementToCompareAgainst
+  );
+}
+
+function renderDiffInternal(
+  elementToBeCompared: Element,
+  elementToCompareAgainst: Element,
   filterToIgnore: DiffFilter<Element> = {},
-  searchElementToBeCompared?: Element,
-  searchElementToCompareAgainst?: Element
+  searchElementToBeCompared: Element,
+  searchElementToCompareAgainst: Element
 ): TemplateResult | null {
   // Determine the ID from the current tag. These can be numbers or strings.
   let idTitle: string | undefined = identity(elementToBeCompared).toString();