+++ test/unit/editors/__snapshots__/SMVSubscriberLaterBinding.test.snap.js
@@ -4,8 +4,13 @@ export const snapshots = {};
 snapshots["SMV Subscribe Later Binding Plugin looks like the latest snapshot"] = 
 `<div>
   <div class="container">
-    <svc-later-binding-list class="column">
-    </svc-later-binding-list>
+    <fcda-later-binding-list
+      class="column"
+      controltag="SampledValueControl"
+    >
+    </fcda-later-binding-list>
+    <extref-later-binding-list class="column">
+    </extref-later-binding-list>
   </div>
 </div>
 `;