+++ test/unit/foundation/__snapshots__/compare.test.snap.js
@@ -1,7 +1,7 @@
 /* @web/test-runner snapshot v1 */
 export const snapshots = {};
 
-snapshots["compas-compare-dialog renderDiff child is added, so check latest snapshot"] = 
+snapshots["compas-compare-dialog renderDiff child is added, so check latest snapshot"] =
 `<div>
   <div class="container container--alt">
     <div class="list__container list__container--left">
@@ -51,7 +51,7 @@ snapshots["compas-compare-dialog renderDiff child is added, so check latest snap
 `;
 /* end snapshot compas-compare-dialog renderDiff child is added, so check latest snapshot */
 
-snapshots["compas-compare-dialog renderDiff child is removed and attribute added/removed/updated, so check latest snapshot"] = 
+snapshots["compas-compare-dialog renderDiff child is removed and attribute added/removed/updated, so check latest snapshot"] =
 `<div>
   <div class="container container--alt">
     <div class="list__container list__container--left">
@@ -223,7 +223,7 @@ snapshots["compas-compare-dialog renderDiff child is removed and attribute added
 `;
 /* end snapshot compas-compare-dialog renderDiff child is removed and attribute added/removed/updated, so check latest snapshot */
 
-snapshots["compas-compare-dialog renderDiff attribute is updated, but is ignored, so check latest snapshot"] = 
+snapshots["compas-compare-dialog renderDiff attribute is updated, but is ignored, so check latest snapshot"] =
 `<div>
   <div class="container container--alt">
     <div class="list__container list__container--left">